pyproject.toml
[tool.poetry]
name = "exploring-langgraph"
version = "0.1.0"
description = "exploring langgraph"
authors = ["Nazareka <nazareka.dev@gmail.com>"]
readme = "README.md"
packages = [{include = "exploring_langgraph"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain-community = "^0.0.13"
langgraph = "^0.0.15"
langchain = "^0.1.1"
python-dotenv = "^1.0.0"
openai = "^1.9.0"
icecream = "^2.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
